added custom hook for cadence gauge
